runCase4 runs correctly now.
Earlier it was bailing out b/c the Fillet operation was failing (because I commented out
the Build command). Turns out there's an error in the opencascade TNaming example - they
try to get back a reference to an Edge that was writen to the Data Framework "only for
example" but it turns out that the Edge that is gotten IsNull. Not sure why, though this
may be an issue for us.